@@ -1,14 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    buildToolsVersion '30.0.3'
+    compileSdkVersion 30
     defaultConfig {
         applicationId "com.hwadzan.fabao"
-        minSdkVersion 19
-        targetSdkVersion 28
+        minSdkVersion 24
+        targetSdkVersion 30
         versionCode 10703
         versionName '1.7.3'
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         ndk {
             //设置支持的SO库架构
             abiFilters 'armeabi', 'armeabi-v7a', 'x86' , 'x86_64', 'arm64-v8a'
@@ -22,33 +23,34 @@ android {
     }
     productFlavors {
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:28.0.0'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    implementation 'com.google.android.material:material:1.3.0'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     //自动升级检测
-    implementation "com.github.yjfnypeu:UpdatePlugin:3.1.2"
+    //implementation "com.github.yjfnypeu:UpdatePlugin:3.1.2"
     //implementation 'com.tencent.bugly:crashreport_upgrade:1.3.6'
     //implementation 'com.tencent.bugly:nativecrashreport:3.6.0.1'
     implementation 'com.qmuiteam:qmui:1.2.0'
-    implementation 'com.squareup.okhttp3:okhttp:3.12.1'
-    implementation 'com.google.code.gson:gson:2.8.5'
-    implementation 'com.liulishuo.okdownload:okdownload:1.0.6'
-    implementation 'com.liulishuo.okdownload:sqlite:1.0.6'
-    implementation 'com.liulishuo.okdownload:okhttp:1.0.6'
-    implementation 'com.github.barteksc:android-pdf-viewer:3.1.0-beta.1'
-    implementation 'com.github.bumptech.glide:glide:4.8.0'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.14.9'
+    implementation 'com.google.code.gson:gson:2.8.7'
+    implementation 'com.github.barteksc:android-pdf-viewer:3.2.0-beta.1'
+    implementation 'com.github.bumptech.glide:glide:4.12.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'//注解处理器
+    implementation 'androidx.recyclerview:recyclerview:1.2.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'org.jetbrains:annotations-java5:15.0'
     //implementation 'com.liulishuo.filedownloader:library:1.7.6'
-    implementation 'com.github.huburt-Hu:NewbieGuide:v2.4.0'
+    implementation 'com.github.huburt-Hu:NewbieGuide:v2.4.4'
 }
