@@ -1,6 +1,6 @@
 package com.hwadzan.ebook.ui;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.hwadzan.ebook.R;
